backend/src/requirements.txt
appdirs==1.4.3
backports.weakref==1.0.post1
bleach==1.5.0
click==6.7
enum34==1.1.6
Flask==0.12
funcsigs==1.0.2
futures==3.2.0
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.9.5
Markdown==2.6.10
MarkupSafe==1.0
mock==2.0.0
numpy==1.13.3
olefile==0.44
packaging==16.8
pbr==3.1.1
Pillow==4.3.0
protobuf==3.5.0.post1
pyparsing==2.2.0
simplejson==3.11.1
six==1.10.0
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
Werkzeug==0.12.1
